+++ utils/spanHTTPUtils.go
@@ -16,6 +16,6 @@ func (s SpanProtocolUtil) AddHTTPSpanProperties() {
 	s.spanDetails.Scheme = GetSpanAttributeValue[string](HTTPSchemeAttrId, s.spanDetailsMap, s.executorAttrStore, s.functionFactory)
 	s.spanDetails.Path = GetSpanAttributeValue[string](HTTPPathAttrId, s.spanDetailsMap, s.executorAttrStore, s.functionFactory)
 	s.spanDetails.Query = GetSpanAttributeValue[string](HTTPQueryAttrId, s.spanDetailsMap, s.executorAttrStore, s.functionFactory)
-	s.spanDetails.Status = GetSpanAttributeValue[int](HTTPStatusAttrId, s.spanDetailsMap, s.executorAttrStore, s.functionFactory)
+	s.spanDetails.Status = GetSpanAttributeValue[float64](HTTPStatusAttrId, s.spanDetailsMap, s.executorAttrStore, s.functionFactory)
 	s.spanDetails.Username = GetSpanAttributeValue[string](HTTPUsernameAttrId, s.spanDetailsMap, s.executorAttrStore, s.functionFactory)
 }
